sidecar/pyproject.toml
[project]
name = "duet-sidecar"
version = "0.1.0"
description = "Duet Python sidecar: transcription, NLP analysis, document parsing"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.1.0",
    "spacy>=3.8.0",
    "PyMuPDF>=1.25.0",
    "python-docx>=1.1.0",
    "anthropic>=0.52.0",
    "librosa>=0.11.0",
    "assemblyai>=0.59.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
duet-sidecar = "duet_sidecar.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
